requirements.txt
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-rq==2.4.1
drf-spectacular==0.20.1
pycryptodome==3.11.0
psycopg2==2.9.1
python-magic-bin==0.4.14
python-libnmap==0.7.2
python-telegram-bot==13.7
pyyaml==5.4.1
requests==2.26.0
